requirements.txt
altgraph==0.17.5
certifi==2026.1.4
charset-normalizer==3.4.4
filelock==3.20.0
fsspec==2025.12.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
onnx==1.20.1
onnxruntime-directml>=1.23.0
opencv-python==4.8.1.78
packaging==26.0
pandas==2.3.3
pefile==2024.8.26
pillow==12.0.0
PyQt5==5.15.10
python-osc==1.8.3
tqdm==4.67.1
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.1
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32-ctypes==0.2.3
requests==2.32.5
six==1.17.0
sympy==1.14.0
# torch is only needed for training, not inference
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
